new BusinessDays functions and setBankHolidays, setHolidays
