+++ .gitattributes
@@ -1,0 +1,8 @@
+/docs               export-ignore
+/tests              export-ignore
+/.gitattributes     export-ignore
+/.gitignore         export-ignore
+/.travis.yml        export-ignore
+/dates.text         export-ignore
+/todo               export-ignore
+/phpunit.xml        export-ignore
