+++ .travis.yml
@@ -5,8 +5,8 @@ env:
 
 matrix:
   include:
-    - php: 5.6
-      env: CARBON_V=1
+    - php: 8.0
+      env: CARBON_V=2
     - php: 7.0
       env: CARBON_V=1
     - php: 7.1
@@ -16,7 +16,7 @@ matrix:
     - php: 7.1
       env: CARBON_V=2
     - php: 7.2
-      env: CARBON_V=2  
+      env: CARBON_V=2
     - php: 7.3
       env: CARBON_V=2
     - php: 7.4
