+++ composer.json
@@ -4,15 +4,16 @@
     "keywords": ["Bank Holiday", "Holiday", "Holidays", "Carbon", "US Holiday", "US Holidays"],
     "license": "MIT",
     "require": {
-        "nesbot/carbon": "^1.22|^2.24.0"
+        "nesbot/carbon": "^1.22|^2.24.0",
+        "php": "^7.0.0 || ^8.0"
     },
     "autoload": {
         "psr-4": {
-            "USHolidays\\": "src/"
+            "USHolidays\\": "src/USHolidays"
         }
     },
     "require-dev": {
-        "phpunit/phpunit": "^5|^6|^7.5|^8.0"
+        "phpunit/phpunit": "^5|^6|^7.5|^8.0|^9.0"
     },
     "config": {
         "discard-changes": true
