+++ src/USHolidays/Traits/Holidays/FlagDay.php
@@ -1,0 +1,28 @@
+<?php
+
+namespace USHolidays\Traits\Holidays;
+
+use USHolidays\Carbon;
+
+trait FlagDay
+{
+    /**
+     * Setting Flag Day
+     *
+     * @param int $year The year to get the holiday in
+     */
+    private function setFlagDay($year)
+    {
+        return Carbon::create($year, 6, 14, 0, 0, 0);
+    }
+
+    /**
+     * Return object of Flag Day for given year
+     *
+     * @param int|null $year The year to get the holiday in
+     */
+    public function getFlagDayHoliday($year = null)
+    {
+        return $this->getHolidaysByYear("Flag Day", $year)[0];
+    }
+}
