+++ src/USHolidays/Traits/Holidays/MLKDay.php
@@ -1,0 +1,34 @@
+<?php
+
+namespace USHolidays\Traits\Holidays;
+
+use USHolidays\Carbon;
+
+trait MLKDay
+{
+    /**
+     * Setting MLK Day
+     *
+     * @param int $year The year to get the holiday in
+     */
+    private function setMLKDay($year)
+    {
+        $date = Carbon::create($year, 1, 1, 0, 0, 0);
+        if( $date->dayOfWeek !== Carbon::MONDAY ) {
+            $date->next(Carbon::MONDAY);
+        }
+        $date->next(Carbon::MONDAY)->next(Carbon::MONDAY);
+
+        return $date;
+    }
+
+    /**
+     * Return object of MLK Day for given year
+     *
+     * @param int|null $year The year to get the holiday in
+     */
+    public function getMLKDayHoliday($year = null)
+    {
+        return $this->getHolidaysByYear("Martin Luther King Jr. Day", $year)[0];
+    }
+}
