+++ src/USHolidays/Traits/Holidays/NewYearsEve.php
@@ -1,0 +1,28 @@
+<?php
+
+namespace USHolidays\Traits\Holidays;
+
+use USHolidays\Carbon;
+
+trait NewYearsEve
+{
+    /**
+     * Setting New Years Eve
+     *
+     * @param int $year The year to get the holiday in
+     */
+    private function setNewYearsEve($year)
+    {
+        return Carbon::create($year, 12, 31, 0, 0, 0);
+    }
+
+    /**
+     * Return object of New Years Eve for given year
+     *
+     * @param int|null $year The year to get the holiday in
+     */
+    public function getNewYearsEveHoliday($year = null)
+    {
+        return $this->getHolidaysByYear("New Year's Eve", $year)[0];
+    }
+}
