+++ src/USHolidays/Traits/Holidays/PalmSunday.php
@@ -1,0 +1,28 @@
+<?php
+
+namespace USHolidays\Traits\Holidays;
+
+use USHolidays\Carbon;
+
+trait PalmSunday
+{
+    /**
+     * Setting Palm Sunday
+     *
+     * @param int $year The year to get the holiday in
+     */
+    private function setPalmSunday($year)
+    {
+        return $this->setEaster($year)->subWeeks(1);
+    }
+
+    /**
+     * Return object of Palm Sunday for given year
+     *
+     * @param int|null $year The year to get the holiday in
+     */
+    public function getPalmSundayHoliday($year = null)
+    {
+        return $this->getHolidaysByYear("Palm Sunday", $year)[0];
+    }
+}
