+++ src/USHolidays/Traits/Holidays/ValentinesDay.php
@@ -1,0 +1,28 @@
+<?php
+
+namespace USHolidays\Traits\Holidays;
+
+use USHolidays\Carbon;
+
+trait ValentinesDay
+{
+    /**
+     * Setting Valentines Day
+     *
+     * @param int $year The year to get the holiday in
+     */
+    private function setValentinesDay($year)
+    {
+        return Carbon::create($year, 2, 14, 0, 0, 0);
+    }
+
+    /**
+     * Return object of Valentines Day for given year
+     *
+     * @param int|null $year The year to get the holiday in
+     */
+    public function getValentinesDayHoliday($year = null)
+    {
+        return $this->getHolidaysByYear("Valentine's Day", $year)[0];
+    }
+}
