+++ src/USHolidays/Traits/Holidays/VeteransDay.php
@@ -1,0 +1,28 @@
+<?php
+
+namespace USHolidays\Traits\Holidays;
+
+use USHolidays\Carbon;
+
+trait VeteransDay
+{
+    /**
+     * Setting Veterans Day
+     *
+     * @param int $year The year to get the holiday in
+     */
+    private function setVeteransDay($year)
+    {
+        return Carbon::create($year, 11, 11, 0, 0, 0);
+    }
+
+    /**
+     * Return object of Veterans Day for given year
+     *
+     * @param int|null $year The year to get the holiday in
+     */
+    public function getVeteransDayHoliday($year = null)
+    {
+        return $this->getHolidaysByYear("Veterans Day", $year)[0];
+    }
+}
