+++ tests/Utilities/NextHolidayTest.php
@@ -15,6 +15,11 @@ class NextHolidayTest extends TestCase
                 ->isSameDay(Carbon::createFromDate(2021, 1, 18))
         );
 
+        $this->assertTrue(
+            $carbon->getNextHolidays(2)[1]->date
+                ->isSameDay(Carbon::createFromDate(2021, 2, 2))
+        );
+
         $this->assertFalse(
             $carbon->getNextHolidays()[0]->date
                 ->isSameDay(Carbon::createFromDate(2021, 1, 1))
