+++ tests/carbon-2.sh
@@ -1,9 +1,11 @@
 #!/bin/bash
 
 cp composer.json original-composer.json
+rm composer.lock
 
+rm -rf vendor
 composer remove nesbot/carbon
-composer require nesbot/carbon:2.*
+composer require nesbot/carbon:2.* --no-plugins
 
 rm composer.json
 mv original-composer.json composer.json
